Use Heroku Config Vars
Implemented config file to avoid having the api keys saved in version control.

Worked locally, but was crashing when I tried to push up to Heroku.

Realized that since the config file wasn't being pushed up to Heroku, it was crashing when told to go grab the config file and load in the api values.

Working Locally Now.